Build bench scripts in native
