+++ ckb-vm-bench-scripts/contracts/ed25519/Cargo.toml
@@ -8,3 +8,11 @@ ckb-std = "0.17.0"
 curve25519-dalek = { version = "4.1.2", default-features = false, features = ["digest"] }
 ed25519-dalek = { version = "2.1.1", default-features = false, features = ["digest"] }
 sha2 = { version = "0.10", default-features = false }
+
+[[bin]]
+name = "ed25519_native"
+path = "src/ed25519_native.rs"
+
+[[bin]]
+name = "ed25519_ckbvm"
+path = "src/ed25519_ckbvm.rs"
