+++ ckb-vm-bench-scripts/contracts/ed25519/Makefile
@@ -35,15 +35,29 @@ endif
 default: build test
 
 build:
-	RUSTFLAGS="$(FULL_RUSTFLAGS)" TARGET_CC="$(CLANG)" TARGET_AR="$(AR)" \
-		cargo build --target=riscv64imac-unknown-none-elf $(MODE_ARGS) $(CARGO_ARGS)
+	for binary in $(BINARIES); do \
+		binary=$${binary}_ckbvm; \
+		RUSTFLAGS="$(FULL_RUSTFLAGS)" TARGET_CC="$(CLANG)" TARGET_AR="$(AR)" \
+			cargo build --target=riscv64imac-unknown-none-elf --bin $$binary $(MODE_ARGS) $(CARGO_ARGS); \
+	done
+	for binary in $(BINARIES); do \
+		binary=$${binary}_native; \
+		TARGET_CC="$(CLANG)" TARGET_AR="$(AR)" \
+			cargo build --bin $$binary $(MODE_ARGS) $(CARGO_ARGS); \
+	done
 	@set -eu; \
 	if [ "x$(BUILD_DIR)" != "x" ]; then \
 		for binary in $(BINARIES); do \
+			binary=$${binary}_ckbvm; \
 			echo "Copying binary $$binary to build directory"; \
 			cp $(TOP)/target/riscv64imac-unknown-none-elf/$(MODE)/$$binary $(TOP)/$(BUILD_DIR); \
 			cp $(TOP)/$(BUILD_DIR)/$$binary $(TOP)/$(BUILD_DIR)/$$binary.debug; \
 			$(OBJCOPY) --strip-debug --strip-all $(TOP)/$(BUILD_DIR)/$$binary; \
+		done; \
+		for binary in $(BINARIES); do \
+			binary=$${binary}_native; \
+			echo "Copying binary $$binary to build directory"; \
+			cp $(TOP)/target/$(MODE)/$$binary $(TOP)/$(BUILD_DIR); \
 		done \
 	fi
 
@@ -64,7 +78,7 @@ fmt:
 # Arbitrary cargo command is supported here. For example:
 #
 # make cargo CARGO_CMD=expand CARGO_ARGS="--ugly"
-# 
+#
 # Invokes:
 # cargo expand --ugly
 CARGO_CMD :=
