+++ ckb-vm-bench-scripts/contracts/ed25519/src/ed25519_ckbvm.rs
@@ -1,0 +1,18 @@
+#![cfg_attr(not(test), no_std)]
+#![cfg_attr(not(test), no_main)]
+
+#[cfg(test)]
+extern crate alloc;
+
+#[cfg(not(test))]
+ckb_std::entry!(program_entry);
+// By default, the following heap configuration is used:
+// * 16KB fixed heap
+// * 1.2MB(rounded up to be 16-byte aligned) dynamic heap
+// * Minimal memory block in dynamic heap is 64 bytes
+// For more details, please refer to ckb-std's default_alloc macro
+// and the buddy-alloc alloc implementation.
+ckb_std::default_alloc!(16384, 1258306, 64);
+
+#[allow(unused_imports)]
+use ed25519::program_entry;
