+++ ckb-vm-bench-scripts/contracts/ed25519/src/ed25519_native.rs
@@ -1,0 +1,5 @@
+use ed25519::program_entry;
+
+fn main() {
+    assert_eq!(program_entry(), 0);
+}
