+++ ckb-vm-bench-scripts/contracts/ed25519/src/lib.rs
@@ -1,0 +1,25 @@
+#![no_std]
+
+use ed25519_dalek::{Signature, VerifyingKey};
+use sha2::{Sha512, digest::Digest};
+
+pub fn program_entry() -> i8 {
+    // Test case from: https://github.com/dalek-cryptography/ed25519-dalek/blob/02001d8c3422fb0314b541fdb09d04760f7ab4ba/tests/ed25519.rs#L102
+    let pub_bytes = [
+        0xec, 0x17, 0x2b, 0x93, 0xad, 0x5e, 0x56, 0x3b, 0xf4, 0x93, 0x2c, 0x70, 0xe1, 0x24, 0x50, 0x34, 0xc3, 0x54,
+        0x67, 0xef, 0x2e, 0xfd, 0x4d, 0x64, 0xeb, 0xf8, 0x19, 0x68, 0x34, 0x67, 0xe2, 0xbf,
+    ];
+    let msg_bytes = [0x61, 0x62, 0x63];
+    let sig_bytes = [
+        0x98, 0xa7, 0x02, 0x22, 0xf0, 0xb8, 0x12, 0x1a, 0xa9, 0xd3, 0x0f, 0x81, 0x3d, 0x68, 0x3f, 0x80, 0x9e, 0x46,
+        0x2b, 0x46, 0x9c, 0x7f, 0xf8, 0x76, 0x39, 0x49, 0x9b, 0xb9, 0x4e, 0x6d, 0xae, 0x41, 0x31, 0xf8, 0x50, 0x42,
+        0x46, 0x3c, 0x2a, 0x35, 0x5a, 0x20, 0x03, 0xd0, 0x62, 0xad, 0xf5, 0xaa, 0xa1, 0x0b, 0x8c, 0x61, 0xe6, 0x36,
+        0x06, 0x2a, 0xaa, 0xd1, 0x1c, 0x2a, 0x26, 0x08, 0x34, 0x06,
+    ];
+    let verifying_key = VerifyingKey::from_bytes(&pub_bytes).unwrap();
+    let sig = Signature::try_from(&sig_bytes[..]).unwrap();
+    let mut prehash_for_verifying = Sha512::default();
+    prehash_for_verifying.update(&msg_bytes[..]);
+    assert!(verifying_key.verify_prehashed(prehash_for_verifying.clone(), None, &sig).is_ok());
+    0
+}
